Revert "Ticket_15_w7Ali Refactored the views and settings and .evn files"
