+++ .gitignore
@@ -7,8 +7,7 @@ __pycache__/
 *.pyc
 db.sqlite3
 /migrations/
-/vcdb/
-/vcdb/db.sqlite3
+
 # Ignore virtual environment
 .venv/
 venv/
@@ -49,10 +48,3 @@ Thumbs.db
 
 # Ignore dummy data
 app/dummy_data
-
-# Ignore static files
-static/
-staticfiles/
-
-# Ignore media files
-media/
