+++ app/urls.py
@@ -4,6 +4,7 @@ from django.contrib.auth import views as auth_views
 from django.urls import path
 
 from app import api_views, views
+
 from .forms import (
     LoginForm,
     MyPasswordChangeForm,
@@ -11,38 +12,27 @@ from .forms import (
     MySetPasswordForm,
 )
 
-# Define URL patterns for the main app
 urlpatterns = [
-    # Product Views
     path("", views.ProductView.as_view(), name="home"),
     path("product-detail/<int:pk>/", views.ProductDetailView.as_view(), name="product-detail"),
-
-    # Cart Actions
     path("add-to-cart/", views.add_to_cart, name="add-to-cart"),
     path("cart/", views.show_cart, name="showcart"),
-    path("pluscart/", views.update_cart, {'action': 'plus'}, name='pluscart'),
-    path("minuscart/", views.update_cart, {'action': 'minus'}, name='minuscart'),
-    path("removecart/", views.remove_cart, name="removecart"),
-
-    # Checkout and Orders
-    path("checkout/", views.checkout, name="checkout"),
-    path("orders/", views.orders, name="orders"),
-    path("paymentdone/", views.payment_done, name="paymentdone"),
-
-    # Profile and Address
-    path("profile/", views.ProfileView.as_view(), name="profile"),
-    path('address/', views.address_view, name='address'),
-
-    # Collection Views
-    path("men_collection/", views.men_collection, name="men_collection"),
-    path("women_collection/", views.women_collection, name="women_collection"),
-
-    # Static Pages
     path("terms-conditions/", views.terms_conditions, name="terms-conditions"),
     path("privacy/", views.privacy, name="privacy"),
+    path("pluscart/", views.plus_cart),
+    path("minuscart/", views.minus_cart),
+    path("removecart/", views.remove_cart),
+    path("checkout/", views.checkout, name="checkout"),
+    path('address/', views.address_view, name='address'),
+    path("orders/", views.orders, name="orders"),
+    path("paymentdone/", views.payment_done, name="paymentdone"),
+    # path("mobile/", views.mobile, name="mobile"),
+    # path("mobile/<slug:data>", views.mobile, name="mobiledata"),
+    path("profile/", views.ProfileView.as_view(), name="profile"),
+    path("men_collection/", views.men_collection, name="men_collection"),
+    path("women_collection/", views.women_collection, name="women_collection"),
 ]
 
-# Define URL patterns for authentication views
 auth_urls = [
     path(
         "accounts/login/",
@@ -101,21 +91,17 @@ auth_urls = [
         views.CustomerRegistrationView.as_view(),
         name="customerregistration",
     ),
-    path("logout/", auth_views.LogoutView.as_view(next_page="login"), name="logout"),
-]
+        path("logout/", auth_views.LogoutView.as_view(next_page="login"), name="logout"),
+] 
 
-# Define URL patterns for API views
 api_urls = [
-    path("api/company/", api_views.company_detail_api, name="company-detail-api"),
-    path("api/products/", api_views.latest_and_discount_product, name="product"),
+    path("api/company/", api_views.company_detail_api, name="company-detail-api"    ),
+    path("api/products", api_views.latest_and_discount_product, name="product"),
     path('api/address/<int:id>/update/', api_views.update_address, name='api-update-address'),
     path('api/address/<int:id>/delete/', api_views.delete_address, name='api-delete-address'),
     path('api/address/set-active/', api_views.set_active_address, name='set-active-address'),
 ]
 
-# Combine all URL patterns
 urlpatterns += auth_urls
 urlpatterns += api_urls
-
-# Serve media files in development
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
